Let regetron work without readline support.
An attempt was made to do this, but unfortunately, it was broken.  If
you don't have readline, it would still attempt to access the readline
module causing an UnboundLocalError to occur.  Put the try/except around
the whole block so that an ImportError causes us to skip doing anything
with readline at all.